+++ app/src/main/java/com/example/swapit/ui/user/ProfileCard.kt
@@ -1,0 +1,91 @@
+package com.example.swapit.ui.user
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.example.swapit.R
+import com.example.swapit.ui.theme.Black
+import com.example.swapit.ui.theme.Gray4
+import com.example.swapit.ui.theme.Paddings
+import com.example.swapit.ui.theme.Typography
+import com.example.swapit.ui.theme.White
+
+@Composable
+fun ProfileCard(
+    userName: String = "하울의움직이는성",
+    userEmail: String = "swapit202501@gmail.com",
+) {
+    Card(
+        modifier =
+            Modifier
+                .fillMaxWidth()
+                .padding(Paddings.xlarge, Paddings.smallMedium),
+        colors =
+            CardDefaults.cardColors(
+                containerColor = White,
+            ),
+        shape = RoundedCornerShape(20.dp),
+    ) {
+        Row(
+            modifier =
+                Modifier
+                    .fillMaxWidth()
+                    .padding(16.dp),
+        ) {
+            Image(
+                painter = painterResource(id = R.drawable.ic_add_plus_circle),
+                contentDescription = "Profile Picture",
+                modifier =
+                    Modifier
+                        .size(86.dp)
+                        .clip(CircleShape),
+            )
+            Spacer(modifier = Modifier.width(16.dp))
+            Column(
+                modifier = Modifier.height(86.dp),
+                verticalArrangement = Arrangement.Center,
+                horizontalAlignment = Alignment.Start,
+            ) {
+                Row(
+                    verticalAlignment = Alignment.CenterVertically,
+                ) {
+                    Text(userName, style = Typography.titleLarge, color = Black)
+                    Spacer(modifier = Modifier.width(8.dp))
+                    Icon(
+                        painter = painterResource(id = R.drawable.ic_pencil),
+                        contentDescription = "프로필 수정",
+                        tint = Gray4,
+                    )
+                }
+                Spacer(modifier = Modifier.height(10.dp))
+                Text(userEmail, style = Typography.bodyMedium, color = Gray4)
+            }
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun ProfileCardPreview() {
+    ProfileCard()
+}
