+++ app/src/main/java/com/example/swapit/ui/user/ProfileItem.kt
@@ -1,0 +1,53 @@
+package com.example.swapit.ui.user
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.example.swapit.R
+import com.example.swapit.ui.theme.Black
+import com.example.swapit.ui.theme.Typography
+
+@Composable
+fun ProfileItem(
+    text: String,
+    count: Int,
+) {
+    Row(
+        modifier =
+            Modifier
+                .fillMaxWidth()
+                .height(56.dp),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        Spacer(modifier = Modifier.width(16.dp))
+        Text(text = text, style = Typography.bodyMedium, color = Black)
+        Spacer(modifier = Modifier.width(6.dp))
+        Text(text = count.toString(), style = Typography.bodyMedium, color = Black)
+        Spacer(modifier = Modifier.weight(1f))
+        Icon(
+            painter = painterResource(id = R.drawable.ic_chevron_right),
+            contentDescription = "더보기",
+            tint = Black,
+        )
+        Spacer(modifier = Modifier.width(16.dp))
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun ProfileItemPreview() {
+    ProfileItem(
+        text = "내가 등록한 물건",
+        count = 5,
+    )
+}
